Rename gh #169 : Renaming the test files to align with their class names
